requirements.txt
aws_lambda_typing
boto3
botocore.exceptions